added distribution
added the source and distribution directories
