+++ distribution/xf.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var Component = require('.lib/core/components/Component');
+
+var xf = {
+  Component: Component
+};
+
+module.exports = xf;
