+++ source/core/Game.js
