+++ source/core/GameObject.js
