+++ source/core/Scene.js
