+++ source/core/components/Component.js
