+++ source/core/components/Identifiable.js
