+++ source/core/managers/Manager.js
