+++ source/core/managers/SceneManager.js
