+++ source/geometry/Polygon.js
