+++ source/geometry/Rectangle.js
