+++ source/graphics/Sprite.js
