+++ source/graphics/components/Animatable.js
